+++ PaperMania/Server/Api/Filter/SessionValidationFilter.cs
@@ -21,7 +21,7 @@ public class SessionValidationFilter : IAsyncActionFilter
         if (!context.HttpContext.Request.Headers.TryGetValue("Session-Id", out var sessionId) || string.IsNullOrWhiteSpace(sessionId))
         {
             _logger.LogWarning("세션 ID가 없습니다.");
-            context.Result = new JsonResult(ApiResponse.Error<EmptyResponse>((int)ErrorStatusCode.Unauthorized, "SID가 없습니다."));
+            context.Result = new JsonResult(ApiResponse.Error<EmptyResponse>(ErrorStatusCode.Unauthorized, "SID가 없습니다."));
             return;
         }
 
@@ -29,7 +29,7 @@ public class SessionValidationFilter : IAsyncActionFilter
         if (!isValid)
         {
             _logger.LogWarning($"유효하지 않은 세션: {sessionId}");
-            context.Result = new JsonResult(ApiResponse.Error<EmptyResponse>((int)ErrorStatusCode.Unauthorized, "유효하지 않은 세션입니다."));
+            context.Result = new JsonResult(ApiResponse.Error<EmptyResponse>(ErrorStatusCode.Unauthorized, "유효하지 않은 세션입니다."));
             return;
         }
